requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
chromadb>=0.5.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
streamlit>=1.36.0
python-dotenv>=1.0.0
groq>=0.9.0
